Исправляет код
