+++ js/data.js
@@ -4,15 +4,6 @@ const MIN_LIKES = 15;
 const MAX_LIKES = 200;
 const COMMENT_COUNT = 10;
 
-const COMMENT_SENTENCES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
 const DESCRIPTIONS = [
   'Фотография солнечного летнего дня. Низкое солнце плывёт среди облаков.',
   'На фотографии новенький спорткар',
@@ -23,8 +14,4 @@ const DESCRIPTIONS = [
   'Дневной поход в ресторан',
 ];
 
-const NAMES = [
-  'Сергей', 'Ольга', 'Валерия', 'Дмитрий', 'Игорь', 'Василий',
-];
-
-export {IMAGE_COUNT, AVATAR_COUNT, MIN_LIKES, MAX_LIKES, COMMENT_COUNT, COMMENT_SENTENCES, DESCRIPTIONS, NAMES};
+export {IMAGE_COUNT, AVATAR_COUNT, MIN_LIKES, MAX_LIKES, COMMENT_COUNT,DESCRIPTIONS,};
